break scenewidgets into classes
