+++ meerk40t/gui/simulation.py
@@ -2,6 +2,7 @@ import math
 
 import wx
 
+from .scene.gridwidget import GridWidget
 from ..core.cutcode import CutCode, LineCut
 from ..kernel import Job, signal_listener
 from ..svgelements import Matrix
@@ -14,7 +15,6 @@ from .icons import (
 from .laserrender import LaserRender
 from .mwindow import MWindow
 from .scene.scene import ScenePanel, Widget
-from .scene.scenewidgets import GridWidget
 from .wxutils import disable_window
 
 _ = wx.GetTranslation
